requirements.txt
python-telegram-bot[job-queue]==20.7
psutil==5.9.6
python-dotenv==1.0.0
